+++ src/texture/ImageConvert.hpp
@@ -9,7 +9,7 @@ namespace ImageConvert {
     bool toRGBA32(std::vector<char>& buffer, const TPL::TPLImageFormat type, uint16_t srcWidth, uint16_t srcHeight, const std::vector<char>& data);
     bool fromRGBA32(std::vector<char>& buffer, const TPL::TPLImageFormat type, uint16_t srcWidth, uint16_t srcHeight, const std::vector<char>& data);
 
-    size_t getImageByteSize(const TPL::TPLImageFormat type, uint16_t width, uint16_t height);
+    uint32_t getImageByteSize(const TPL::TPLImageFormat type, uint16_t width, uint16_t height);
 }
 
 #endif // IMAGECONVERT_HPP
